Refactor time stuff out of TemporalInterrupt
